fix annotation ID

+++ backend/src/main/java/com/example/integradorvelosabackend/Repository/AnnotationRepository.java
@@ -3,5 +3,5 @@ package com.example.integradorvelosabackend.Repository;
 import com.example.integradorvelosabackend.Entity.Annotation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AnnotationRepository  extends CrudRepository<Annotation, Integer> {
+public interface AnnotationRepository  extends CrudRepository<Annotation, Long> {
 }
